Initial Set up for Shortest Job First
Changed our designed a little bit. We are not making a deep copy instead
of shallow copy of  the initial Processes ArrayList. Reason for this
change is because if we are making any changes on the Process object in
First come First Served Algorithm, those changes will persist in later
version of the Processes.